backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
Pillow
gtts
